requirements.txt
# Exchange
ccxt>=4.0.0

# Data & Analysis
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0

# Claude AI Sentiment
anthropic>=0.40.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
schedule>=1.2.0
rich>=13.0.0
